% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axMap.R
\name{axMap}
\alias{axMap}
\title{Transformation based on two axis}
\usage{
axMap(input, ax1, ax2, direction)
}
\arguments{
\item{input}{The values you want to transform}

\item{ax1}{First axis limits}

\item{ax2}{Second axis limits}

\item{direction}{Either 'transformAxis' or 'transformValues'}
}
\value{
The trnsformed vlues
}
\description{
Transformation based on two axis
}
\examples{
direction <- 'transformAxis'
input <-  c(800, 900, 1000)
ax1 <- c(80,100)
ax2 <- c(700, 1300)
axMap(input, ax1, ax2, direction)
}
