% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heizPack-package.R
\docType{package}
\name{heizPack-package}
\alias{heizPack}
\alias{heizPack-package}
\title{heizPack: Grafische Darstellung der Heizungswechsel basierend auf Energiemeldungen}
\description{
Zeitreihen des Heizungswechsels von 2019 bis heute.
}
\author{
\strong{Maintainer}: Tobias Sommer \email{tobias.sommer@lu.ch} (\href{https://orcid.org/0000-0001-5274-1976}{ORCID})

}
\keyword{internal}
