% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyEnergiemeldungen.R
\name{sankeyEnergiemeldungen}
\alias{sankeyEnergiemeldungen}
\title{Create Sankey Plots based on dataset Energiemeldungen}
\usage{
sankeyEnergiemeldungen(ejahr)
}
\arguments{
\item{ejahr}{Jahr der Auswertung (integer zwischen 2019 und 2023)}
}
\value{
Grafik der Auswertung
}
\description{
Create Sankey Plots based on dataset Energiemeldungen
}
\examples{
\dontrun{
sankeyEnergiemeldungen(2019)
}
}
