% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyEnergiemeldungen.R
\name{sankeyEnergiemeldungen}
\alias{sankeyEnergiemeldungen}
\title{Create Sankey Plots based on dataset Energiemeldungen}
\usage{
sankeyEnergiemeldungen()
}
\value{
A table
}
\description{
Create Sankey Plots based on dataset Energiemeldungen
}
\examples{
\dontrun{
sankeyEnergiemeldungen()
}
}
